Fix an issue where headers were not being included if they were for a platform
This is when the headers are set for just a platform, they were being excluded
because the if statement to include all public headers was actually checking
something else (source_files).

This would also produce another bug if there was only private headers in a spec
or if there was only implementation files.