[Docset Fixer] Fix issue retrieving latest version
This is due to looking in the local file system for the versions, since
we are now parsing documentation directly from the hook from trunk and
using the JSON from the hook. We no longer have it on disk anymore
because we don't sync the spec repository.

Fixes #228
Fixes #237
Fixes #245
Fixes #246